requirements.txt
Flask==3.1.3
pymupdf==1.27.1
python-docx==1.2.0
werkzeug>=3.1.0